Move endpoints to dedicated files
