+++ src/server/tests/zgloszenie.test.ts
@@ -1,8 +1,8 @@
 import request from "supertest";
-import { ZgloszeniePayload } from "../common/zgloszenieSchema";
-import { FieldValidationError, ValidationErrorBody } from "./middleware/zodValidation";
-
-import app from "./app";
+import { ZgloszeniePayload } from "../../common/zgloszenieSchema";
+import { FieldValidationError, ValidationErrorBody } from "../middleware/zodValidation";
+import app from "../app";
+import "../endpoints/zgloszeneEndpoints"
 
 describe("Dodawanie Zgloszenia - Testy", () => {
